+++ src/constants.ts
@@ -1,7 +1,13 @@
-import { COUNTRIES, GERMANY_COUNTRY, SPAIN_COUNTRY } from "./geo/countries.js";
+import {
+  COUNTRIES,
+  FRANCE_COUNTRY,
+  GERMANY_COUNTRY,
+  SPAIN_COUNTRY,
+} from "./geo/countries.js";
 
 export const SPAIN_CENTER: [number, number] = SPAIN_COUNTRY.center;
 export const GERMANY_CENTER: [number, number] = GERMANY_COUNTRY.center;
+export const FRANCE_CENTER: [number, number] = FRANCE_COUNTRY.center;
 export const CZECHIA_CENTER: [number, number] = COUNTRIES.CZ.center;
 export const POLAND_CENTER: [number, number] = COUNTRIES.PL.center;
 export const DEFAULT_ZOOM = 6;